+++ api/bootstrap/bootstrap_test.go
@@ -7,9 +7,9 @@ import (
 // TODO: Write a much better and more comprehensive test
 func TestBootstrap(t *testing.T) {
 	outDir := "sample_project"
-	modelName := "AwesomeModel"
+	blockName := "AwesomeBlock"
 	defer deleteDir(outDir)
-	err := CreateProject(outDir, modelName)
+	err := CreateProject(outDir, blockName)
 	if err != nil {
 		t.Errorf("Test failed: %s", err.Error())
 	}