+++ api/bootstrap/templates/full/src/block.py
@@ -1,13 +1,13 @@
 from typing import Any, Dict
 
-from clay.core import ModelWrapper
+from clay.core import BlockWrapper
 import datatypes
 import rasterio
 import numpy
 
-class {{.ModelName}}(ModelWrapper):
+class {{.BlockName}}(BlockWrapper):
     def setup(self, weight: int, **hyperparameters) -> None:  # type: ignore
-        # download weights, initialize model,
+        # download weights, initialize block,
         # input name in the function needs to match the parameter name in the spec file
         # setup directories, etc.
         self.weight = weight
@@ -17,7 +17,7 @@ class {{.ModelName}}(ModelWrapper):
         input1: datatypes.String,
         input2: datatypes.Number,
     ) -> Dict[str, Any]:
-        # function takes inputs for a model
+        # function takes inputs for a block
         # input name in the function needs to match input name from spec file
         self.logger.warning(
             "In pre-process. Use self.logger for all logging. Avoid print statements"
@@ -28,7 +28,7 @@ class {{.ModelName}}(ModelWrapper):
     async def inference(self, input1: String, input2: Number) -> Dict[str, Any]:  # type: ignore
         # simply run inference and return the results
         # and anything extra if required
-        self.logger.info("In inference. I can access all `self` parameters throughout the model ")
+        self.logger.info("In inference. I can access all `self` parameters throughout the block ")
         return {"input1": input1, "input2": input2}
 
     async def postprocess(self, input1: String, input2: Number) -> Dict[str, Data]:  # type: ignore