+++ python/clay/run.py
@@ -1,21 +1,21 @@
 import os
 from typing import Type
 
-from clay.core import ModelWrapper
+from clay.core import BlockWrapper
 from clay.runners.runner import JobRunner
 
 
-def Run(model: Type[ModelWrapper], name: str, cfg_path: str) -> None:
-    """The general method to execute a model. In most cases, users and programs alike
-    should be using this method to run the model. The function looks at a bunch of internal
+def Run(block: Type[BlockWrapper], name: str, cfg_path: str) -> None:
+    """The general method to execute a block. In most cases, users and programs alike
+    should be using this method to run the block. The function looks at a bunch of internal
     environment variables and infers which runner *or execution mode* to use.
-    This ensures that the execution mode of the model is abstracted away from the user.
+    This ensures that the execution mode of the block is abstracted away from the user.
 
     Args:
-        model (Type[ModelWrapper]):
-            The user defined model that subclasses `ModelWrapper`.
+        block (Type[BlockWrapper]):
+            The user defined block that subclasses `BlockWrapper`.
         name (str):
-            Name of the model to be run. This is used as an identifier in the in-built logger.
+            Name of the block to be run. This is used as an identifier in the in-built logger.
         cfg_path (str): Path to the configuration that is meant to be used.
 
     Raises:
@@ -28,9 +28,9 @@ def Run(model: Type[ModelWrapper], name: str, cfg_path: str) -> None:
     if not os.path.exists(cfg_path):
         raise FileNotFoundError(cfg_path)
     runner = JobRunner(
-        model_name=name,
-        model_class=model,
-        model_args={"config": cfg_path},
+        block_name=name,
+        block_class=block,
+        block_args={"config": cfg_path},
         cfg_path=cfg_path,
     )
     return runner.start()