+++ python/tests/test_core.py
@@ -5,25 +5,25 @@ from typing import Any
 
 import pytest
 
-from clay import ModelWrapper
+from clay import BlockWrapper
 from clay.core import BaseRunner
 
-from .models.ymxplusc import YMXPLUSC, YMXPLUSC_CONFIG
+from .blocks.ymxplusc import YMXPLUSC, YMXPLUSC_CONFIG
 
 
 def test_mw_missing_setup_override() -> None:
-    class M(ModelWrapper):
+    class M(BlockWrapper):
         async def preprocess(self, *args: Any, **kwargs: Any) -> Any:
             pass
 
     with pytest.raises(NotImplementedError):
-        M(config="./tests/models/ymxplusc.yaml")
+        M(config="./tests/blocks/ymxplusc.yaml")
 
 
 def test_mw_blocking_method_override() -> None:
     with pytest.raises(AssertionError):
 
-        class M(ModelWrapper):
+        class M(BlockWrapper):
             def setup(self) -> None:
                 pass
 
@@ -36,8 +36,8 @@ class TestBaseRunner(unittest.TestCase):
         class DemoRunner(BaseRunner):
             def __init__(self):
                 super().__init__(
-                    modelcls=YMXPLUSC,
-                    model_args={"config": YMXPLUSC_CONFIG},
+                    blockcls=YMXPLUSC,
+                    block_args={"config": YMXPLUSC_CONFIG},
                     logger=None,
                     cfg_path=YMXPLUSC_CONFIG,
                 )
@@ -59,25 +59,25 @@ class TestBaseRunner(unittest.TestCase):
                 new_progress = self._current_progress + progress_delta
                 self.set_progress(new_progress)
 
-        self._test_modelcls = YMXPLUSC
+        self._test_blockcls = YMXPLUSC
         self._test_runnercls = DemoRunner
         time.sleep(1)
 
     def test_setup(self):
         os.environ["SAMPLE_ENV"] = "alreadyExists"
         self.r = self._test_runnercls()
-        self.r._init_model()        
+        self.r._init_block()        
         assert os.environ["SAMPLE_ENV"] == "alreadyExists"
         assert os.environ["SAMPLE_ENV_1"] == "2"
 
-    def test_model_init(self):
+    def test_block_init(self):
         self.r = self._test_runnercls()
-        self.r._init_model()
-        assert isinstance(self.r._model, self._test_modelcls)
+        self.r._init_block()
+        assert isinstance(self.r._block, self._test_blockcls)
 
     def test_progress_update(self):
         self.r = self._test_runnercls()
-        self.r._init_model()
+        self.r._init_block()
 
         self.r.set_progress(13)
         assert self.r.get_progress() == 13
@@ -93,7 +93,7 @@ class TestBaseRunner(unittest.TestCase):
 
     def test_progress_update_with_multiple_increments_greater_than_max(self):
         self.r = self._test_runnercls()
-        self.r._init_model()
+        self.r._init_block()
 
         self.r.set_progress(100)
         assert self.r.get_progress() == 100
@@ -103,7 +103,7 @@ class TestBaseRunner(unittest.TestCase):
 
     def test_progress_update_with_negative_increment(self):
         self.r = self._test_runnercls()
-        self.r._init_model()
+        self.r._init_block()
 
         assert self.r.get_progress() == 0
         self.r.set_progress(-10)