+++ python/tests/test_types.py
@@ -4,12 +4,12 @@
 
 
 # def test_clb_init() -> None:
-#     c = types.Callback(Id="123", State=types.ModelStates.FAILED)
-#     assert c.State == types.ModelStates.FAILED
+#     c = types.Callback(Id="123", State=types.BlockStates.FAILED)
+#     assert c.State == types.BlockStates.FAILED
 
 
 # def test_clb_serialise_dict() -> None:
-#     c = types.Callback(Id="task123", State=types.ModelStates.FAILED, FailureType="bad_request")
+#     c = types.Callback(Id="task123", State=types.BlockStates.FAILED, FailureType="bad_request")
 #     d = c.model_dump(by_alias=True)
 #     assert d == {
 #         "id": "task123",
@@ -30,7 +30,7 @@
 #     }
 
 
-# def test_init_raster_model_from_dict_success() -> None:
+# def test_init_raster_block_from_dict_success() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -40,12 +40,12 @@
 #         "properties": {"bands": ["B01", "B02"]},
 #     }
 
-#     r = types.Raster.model_validate(d, context={"a": "b"})
+#     r = types.Raster.block_validate(d, context={"a": "b"})
 #     assert r.Properties is not None
 #     assert r.Properties.Bands == ["B01", "B02"]
 
 
-# def test_init_raster_model_from_dict_with_none_props_success() -> None:
+# def test_init_raster_block_from_dict_with_none_props_success() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -55,11 +55,11 @@
 #         "properties": None,
 #     }
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r.Properties is None
 
 
-# def test_init_raster_model_from_dict_with_sun_elevation() -> None:
+# def test_init_raster_block_from_dict_with_sun_elevation() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -71,13 +71,13 @@
 #         },
 #     }
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r.Properties is not None
 #     assert r.Properties.SunElevation == 1.2
 #     assert r.Properties.SatelliteLookAngle is None
 
 
-# def test_init_raster_model_from_dict_with_date_in_props() -> None:
+# def test_init_raster_block_from_dict_with_date_in_props() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -89,13 +89,13 @@
 #         },
 #     }
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r.Properties is not None
 #     assert r.Properties.Date == "1/1/2022"
 #     assert r.Properties.SatelliteLookAngle is None
 
 
-# def test_init_raster_model_from_dict_with_satellite_look_angle() -> None:
+# def test_init_raster_block_from_dict_with_satellite_look_angle() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -107,38 +107,38 @@
 #         },
 #     }
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r.Properties is not None
 #     assert r.Properties.SatelliteLookAngle == 3.4
 #     assert r.Properties.SunElevation is None
 
 
-# def test_init_string_model_with_default_value() -> None:
+# def test_init_string_block_with_default_value() -> None:
 #     d = {"format": "string", "type": "str", "name": "s", "default": "hello"}
-#     s = types.String.model_validate(d)
+#     s = types.String.block_validate(d)
 #     assert s.Default == "hello"
 
 
-# def test_init_string_model_with_no_value_or_default() -> None:
+# def test_init_string_block_with_no_value_or_default() -> None:
 #     d = {"format": "string", "type": "str", "name": "s"}
-#     s = types.String.model_validate(d)
+#     s = types.String.block_validate(d)
 #     assert s.Value is None
 #     assert s.Default is None
 
 
-# def test_init_date_model_with_value() -> None:
+# def test_init_date_block_with_value() -> None:
 #     d = {"format": "date", "type": "str", "name": "d", "value": "12-02-2022"}
-#     v = types.Date.model_validate(d)
+#     v = types.Date.block_validate(d)
 #     assert v.Value == "12-02-2022"
 
 
-# def test_init_tabular_model_with_value() -> None:
+# def test_init_tabular_block_with_value() -> None:
 #     d = {"format": "tabular", "type": "url", "name": "t", "value": "f.csv"}
-#     v = types.Tabular.model_validate(d)
+#     v = types.Tabular.block_validate(d)
 #     assert v.Value == "f.csv"
 
 
-# def test_init_raster_model_from_dict_with_continuous_viz() -> None:
+# def test_init_raster_block_from_dict_with_continuous_viz() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -150,7 +150,7 @@
 #         },
 #     }
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r.Properties is not None
 #     assert r.Properties.Visualisation is not None
 #     assert r.Properties.Visualisation.Type == "continuous"
@@ -161,7 +161,7 @@
 #     assert r.Properties.Visualisation.Continuous.Range == [[-1, -2]]
 
 
-# def test_init_raster_model_from_dict_with_discrete_viz() -> None:
+# def test_init_raster_block_from_dict_with_discrete_viz() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -179,7 +179,7 @@
 #         },
 #     }
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r.Properties is not None
 #     assert r.Properties.Visualisation is not None
 #     assert r.Properties.Visualisation.Type == "discrete"
@@ -190,7 +190,7 @@
 #     assert r.Properties.Visualisation.Discrete["2"] == "#colorcode2"
 
 
-# def test_init_raster_model_from_dict_with_bucket_viz() -> None:
+# def test_init_raster_block_from_dict_with_bucket_viz() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -216,7 +216,7 @@
 #         },
 #     }
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r.Properties is not None
 #     assert r.Properties.Visualisation is not None
 #     assert r.Properties.Visualisation.Type == "bucket"
@@ -229,7 +229,7 @@
 #     assert r.Properties.Visualisation.Bucket[0][1].Range == [-5, -4]
 
 
-# def test_init_raster_model_from_dict_with_discretization_interval() -> None:
+# def test_init_raster_block_from_dict_with_discretization_interval() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -270,11 +270,11 @@
 #         ),
 #     )
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r == target
 
 
-# def test_init_raster_model_from_dict_with_discretization_index() -> None:
+# def test_init_raster_block_from_dict_with_discretization_index() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -307,11 +307,11 @@
 #         ),
 #     )
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r == target
 
 
-# def test_init_raster_model_from_dict_with_images_list() -> None:
+# def test_init_raster_block_from_dict_with_images_list() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -330,11 +330,11 @@
 #         ),
 #     )
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r == target
 
 
-# def test_init_raster_model_from_dict_with_group() -> None:
+# def test_init_raster_block_from_dict_with_group() -> None:
 #     d = {
 #         "format": "raster",
 #         "type": "url",
@@ -355,5 +355,5 @@
 #         ),
 #     )
 
-#     r = types.Raster.model_validate(d)
+#     r = types.Raster.block_validate(d)
 #     assert r == target