+++ python/CHANGELOG.md
@@ -2,8 +2,8 @@
 ## v1.3.1a2 - 02-05-2025
 ### New Features
 - Introduced stac_url in Raster datatype
-- Added add_asset function to upload an inout/output asset explicitly from model
-- Added set disclaimer function to be used from model to send out model disclaimers
+- Added add_asset function to upload an inout/output asset explicitly from block
+- Added set disclaimer function to be used from block to send out block disclaimers
 - Removed orchestrator related references from clay
 - Released clay python SDK version 1.3.1a2
 ### Fixes
@@ -42,4 +42,4 @@
 - Added `date` to `RasterProperties` by @sanafirdaus. [(6e50abd)](https://github.com/example/clay/commit/6e50abd875d95b5e886af307a1e8f26cd44ec078)
 - Added time metrics to inference callback by @pranjaldatta. [(c3c192b)](https://github.com/example/clay/commit/c3c192b16f3863453c0580bf8b83c767143f97b0)
 ### Fixes
-- Improved time calculation for total model runtime and inference runtime by @pranjaldatta. [(d2c751)](https://github.com/example/clay/commit/d2c751086cc6fe8deea73f266804a315bf3ed9f0)
+- Improved time calculation for total block runtime and inference runtime by @pranjaldatta. [(d2c751)](https://github.com/example/clay/commit/d2c751086cc6fe8deea73f266804a315bf3ed9f0)